Move test_helper into main tree so it can be used to test end user filesystems
