+++ spec/spec_helper.rb
@@ -2,7 +2,7 @@
 
 require 'minitest/spec'
 require 'minitest/autorun'
-require "minitest/reporters"
+require 'minitest/reporters'
 Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new unless ENV.key?('RM_INFO')
 
 module Enumerable
